Introduce keyboard shortcut
Register a keyboard shortcut (keybinding) to toggle desaturation. This
keybinding is stored using GSettings, just like GNOME keybindings. As
such, it can also be overridden by users to suit their needs, or
disabled entirely.

The keybinding is registered with the window manager, so the extension
doesn't have to do any interpretation of modifier keys, etc. GNOME takes
care of all of that!

The one drawback of this change is that it requires additional steps
during installation in order to compile the GSettings schema and make
~/.local/share a path that is searched for compiled GSettings schemas.
I'm not sure if there's a way around that; let me know if you think
there's a better way to do this.