Introduce keyboard shortcut (#5)
* Introduce keyboard shortcut

Register a keyboard shortcut (keybinding) to toggle desaturation. This
keybinding is stored using GSettings, just like GNOME keybindings. As
such, it can also be overridden by users to suit their needs, or
disabled entirely.

The keybinding is registered with the window manager, so the extension
doesn't have to do any interpretation of modifier keys, etc. GNOME takes
care of all of that!

It turns out that there is an established convention for how GNOME Shell
extensions should ship their GSettings schemas, so that people using the
extension don't have to run any additional installation steps.

Given that this extension doesn't have a preferences/settings UI yet,
I've updated the README.md with instructions on how to change the
default keyboard shortcut using the `gsettings` command-line tool.